factor out engine into its own file
